// src/types/docket.ts
export type DocketT = {
  docketId: string;
  customerName: string;
  customerAddress: string;
  customerEmail: string;
  customerStreet: string;
  customerCity: string;
  customerCounty: string;
  customerEircode: string;
  customerCountry: string;
  jobId: string;
  customerId: string;
  individualDocketNumber: string;
  prefix: string;
  docketNumber: number;
  gpsOn: boolean;
  longitude: string;
  latitude: string;
  date: string;
  time: string;
  vehicleRegistration: string;
  generalPickupDescription: string;
  nonWasteLoadPictures: string[];
  isWaste: boolean;
  wasteDescription: string;
  wasteLoWCode: string;
  isHazardous: boolean;
  localAuthorityOfOrigin: string;
  wasteQuantity: {
    __typename: string;
    unit: string;
    amount: number;
  };
  collectedFromWasteFacility: boolean;
  collectionPointName: string;
  collectionPointAddress: string;
  collectionPointStreet: string;
  collectionPointCity: string;
  collectionPointCounty: string;
  collectionPointEircode: string;
  collectionPointCountry: string;
  destinationFacilityLatitude: string;
  destinationFacilityLongitude: string;
  destinationFacilityName: string;
  destinationFacilityAuthorisationNumber: string;
  destinationFacilityAddress: string;
  destinationFacilityStreet: string;
  destinationFacilityCity: string;
  destinationFacilityCounty: string;
  destinationFacilityEircode: string;
  destinationFacilityCountry: string;
  driverSignature: string;
  wasteFacilityRepSignature: string;
  customerSignature: string;
  isLoadForExport: boolean;
  portOfExport: string;
  countryOfDestination: string;
  facilityAtDestination: string;
  tfsReferenceNumber: string;
  additionalInformation: string;
  creatorEmail: string;
  doSendEmail: string;
  createdAt: string;
};

export type DocketDataT = {
  docketId: string;
  customerName: string;
  customerAddress: string;
  customerEmail: string;
  customerStreet: string;
  customerCity: string;
  customerCounty: string;
  customerEircode: string;
  customerCountry: string;
  jobId: string;
  customerId: string;
  individualDocketNumber: string;
  prefix: string;
  docketNumber: number;
  gpsOn: boolean;
  longitude: string;
  latitude: string;
  date: string;
  time: string;
  vehicleRegistration: string;
  generalPickupDescription: string;
  nonWasteLoadPictures: string[];
  isWaste: boolean;
  wastes: Array<{
    wasteDescription: string;
    wasteLoWCode: string;
    isHazardous: boolean;
    localAuthorityOfOrigin: string;
    wasteQuantity: {
      __typename: string;
      unit: string;
      amount: number;
    };
    wasteLoadPicture: string;
  }>;
  collectedFromWasteFacility: boolean;
  collectionPointName: string;
  collectionPointAddress: string;
  collectionPointStreet: string;
  collectionPointCity: string;
  collectionPointCounty: string;
  collectionPointEircode: string;
  collectionPointCountry: string;
  destinationFacilityLatitude: string;
  destinationFacilityLongitude: string;
  destinationFacilityName: string;
  destinationFacilityAuthorisationNumber: string;
  destinationFacilityAddress: string;
  destinationFacilityStreet: string;
  destinationFacilityCity: string;
  destinationFacilityCounty: string;
  destinationFacilityEircode: string;
  destinationFacilityCountry: string;
  driverSignature: string;
  wasteFacilityRepSignature: string;
  customerSignature: string;
  isLoadForExport: boolean;
  portOfExport: string;
  countryOfDestination: string;
  facilityAtDestination: string;
  tfsReferenceNumber: string;
  additionalInformation: string;
  creatorEmail: string;
  doSendEmail: string;
};